Clean up directory structure a little.
This is an application and not a library, so none of the golang directories need to be
exported.  This will also make it a little easier to see the structure of the app when
looking at it for the first time.  If there's any go code that might be useful outside
of this application, it can be pulled into a separate library at some point.