Allow searching within topics
The search box at the top of the page now supports searching within topics. A topic is
added to the search box with the trigger "in:", and then autocomplete is used to select
the topic. When a topic is added to the search box, results are limited to links and
topics below that topic (including the topic itself). I.e., a search is done within
the transitive closure of the topics beneath the one provided.

If several topics are provided, an intersection is done on the several transitive closures,
and results are limited to topics in the remaining set.