Use BTreeSet and BTreeHash in the Git serializations
There's two things that we can get by using BTreeSet and BTreeHash in the serializable
structs used in the Git module.  First, it will ensure that duplicates are avoided.
And second, it will impose an ordering, which will be important for a text-based
database that makes use of Git's strategy of using diffs to maintain history.